requirements.txt
asgiref==3.11.0
attrs==25.4.0
Django==6.0
djangorestframework==3.16.1
drf-spectacular==0.29.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
psycopg2==2.9.11
python-dotenv==1.2.1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.4
tzdata==2025.2
uritemplate==4.2.0
